pyproject.toml
[tool.poetry]
name = "django-realworld"
version = "0.1.0"
description = ""
authors = ["sawirricardo <sawir.ricardo@gmail.com>"]
readme = "README.md"
packages = [{include = "django_realworld"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
psycopg2 = "^2.9.5"
django-environ = "^0.9.0"
django-extensions = "^3.2.1"
gunicorn = "^20.1.0"


[tool.poetry.group.dev.dependencies]
werkzeug = "^2.2.3"
factory-boy = "^3.2.1"
django-debug-toolbar = "^3.8.1"
black = "^23.1.0"
isort = "^5.12.0"
pytest = "^7.2.1"
pytest-django = "^4.5.2"
django-test-plus = "^2.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
